adds searchBar component, removes search bar from navbar and adds link to SearchPage to allow user search
