+++ frontend/src/common/components/Navbar/Navbar.css
@@ -9,7 +9,7 @@
   align-items: center;
   padding: 30px 40px;
   z-index:1000;
-  opacity: .99;
+  opacity: .5;
 }
 
 .logo {
