+++ frontend/src/index.js
@@ -9,6 +9,7 @@ import HomePage from './home/containers/HomePage/HomePage';
 import LandingPage from './landing/containers/LandingPage/LandingPage';
 import AuthPage from './auth/containers/Authpage/AuthPage';
 import RecipeFormPage from './recipes/containers/RecipeFormPage/RecipeFormPage';
+import SearchPage from './search/containers/SearchPage/SearchPage'
 
 
 const router = createBrowserRouter([
@@ -17,7 +18,8 @@ const router = createBrowserRouter([
     { path: "/recipe", element: <RecipeDetailPage/>},
     { path: "/home", element: <HomePage/>},
     { path: "/login", element: <AuthPage/>},
-    { path: "/create-recipe", element: <RecipeFormPage/>}
+    { path: "/create-recipe", element: <RecipeFormPage/>},
+    {path: "/search", element: <SearchPage/>}
   ]);
   
   ReactDOM.createRoot(document.getElementById("root")).render(
