+++ frontend/src/search/containers/SearchPage/SearchPage.css
@@ -1,0 +1,3 @@
+#search-page-container{
+    margin-top:1011px;
+}
