+++ frontend/src/search/containers/SearchResultPage/SearchResultPage.css
@@ -1,0 +1,3 @@
+#search-result-page-container{
+    margin-top:100px;
+}
