+++ frontend/src/search/containers/SearchResultPage/SearchResultPage.js
@@ -1,9 +1,23 @@
-import React from 'react'
+import React, {useState} from 'react'
 import './SearchResultPage.css'
+import FilterOptions from '../../components/FilterOptions/FilterOptions'
+import Navbar from '../../../common/components/Navbar/Navbar'
+import Footer from '../../../common/components/Footer/Footer'
 
 const SearchResultPage = () => {
+
+  const[showFilter, setShowFilter] = useState(false)
+
+  const toggleFilter = () => {
+    setShowFilter(prevFilter => !prevFilter)
+  }
   return (
-    <div>SearchResultPage</div>
+    <div id="search-result-page-container">
+      <Navbar/>
+      <button onClick ={toggleFilter}>Filter Search</button>
+      {showFilter ? <FilterOptions/> : ""}
+      <Footer/>
+    </div>
   )
 }
